Add +requestForUserWithID: to user requests
With the new caching implementation, we don't actually have a userID
until we complete the request to get the user. This means that the API
client doesn't have a userID to use for the current user request. By
adding a request to get a user with a specific ID, and forwarding the
current user requests to that method, we can use the ID returned with
the authentication request instead of having to pull it out of the API
client.