Bug fix in cached location search
The old search used increasing bounding boxes until it found more
locations than the limit. The problem is that for a given radius, using
a square inscribing will include locations outside of the circle. This
meant that for a given limit, the results could be missing locations
that were actually closer than some of the results.

The `LUCachedLocationSearch` now filters out locations outside of the
circle when doing its bounding box searches. This performs a little
worse because the locations have to be loaded into memory and filtered
each time, but the results will be more accurate.