Add "code" property to LUAPIError
This property can be nil, but when present it represents a
machine-readable identifier for the error. For example, if the error
message is "First name is blank", the code might be "blank".

This makes it easier to look for specific errors without depending on
the exact text of the error message.