Add support for navigation interstitials
Navigation interstitials are much like URL interstitials, but the URL
represents a screen in the app to which the user can navigate (for
example, to buy a gift card).

Also adds "buttonText" as an action property for both navigation and URL
interstitials. This represents the text that should be displayed inside
the button that opens the URL.